improve search components
