+++ .changeset/gold-dingos-breathe.md
@@ -1,0 +1,7 @@
+---
+'pliny': patch
+---
+
+Improve kbar styling
+Simplify kbar component
+Add new KBarButton and AlgoliaButton wrapper which toggles their respective modals on click event
