+++ packages/pliny/package.json
@@ -32,7 +32,7 @@
     "github-slugger": "^1.4.0",
     "image-size": "1.0.0",
     "js-yaml": "4.1.0",
-    "kbar": "0.1.0-beta.40",
+    "kbar": "0.1.0-beta.41",
     "next-contentlayer": "^0.3.4",
     "next-themes": "^0.2.1",
     "remark": "^14.0.2",
