+++ packages/pliny/src/search/KBarButton.tsx
@@ -1,0 +1,20 @@
+import { DetailedHTMLProps, HTMLAttributes } from 'react'
+import { useKBar } from 'kbar'
+
+/**
+ * Button wrapper component that triggers the KBar modal on click.
+ *
+ * @return {*}
+ */
+export const KBarButton = ({
+  children,
+  ...rest
+}: DetailedHTMLProps<HTMLAttributes<HTMLButtonElement>, HTMLButtonElement>) => {
+  const { query } = useKBar()
+
+  return (
+    <button {...rest} onClick={() => query.toggle()}>
+      {children}
+    </button>
+  )
+}
